formated properly to follow project structure
