+++ apps/web/src/types/news.types.ts
