+++ apps/web/src/utils/news.ts
@@ -1,22 +1,7 @@
-import { Article } from "./types";
-import { sanityQuery } from "./sanity";
+import { Article } from "@/types/news.types";
+import { sanityQuery } from "../app/(public)/news/sanity";
 import { endpoints } from "@/data/endpoints";
-
-const getDaysAgo = (value: number) => {
-  const msPerDay = 1000 * 60 * 60 * 24;
-  return Math.floor((Date.now() - new Date(value).getTime()) / msPerDay);
-};
-
-export const matchesTimeRange = (articleDate: number, range: string) => {
-  if (range === "all") return true;
-  const age = getDaysAgo(articleDate);
-  if (range === "last-7") return age <= 7;
-  if (range === "last-30") return age <= 30;
-  if (range === "last-90") return age <= 90;
-  if (range === "this-year")
-    return new Date(articleDate).getFullYear() === new Date().getFullYear();
-  return true;
-};
+import { matchesTimeRange } from "@/utils/date";
 
 export const getFilteredArticles = (
   browseableArticles: Article[],
@@ -32,13 +17,6 @@ export const getFilteredArticles = (
       return matchesField && matchesRegion && matchesTime;
     })
     .sort((l, r) => new Date(r.date).getTime() - new Date(l.date).getTime());
-
-export const formatDate = (value: number) =>
-  new Intl.DateTimeFormat("en-US", {
-    month: "short",
-    day: "numeric",
-    year: "numeric",
-  }).format(new Date(value));
 
 export const escapeHtml = (value: string) =>
   String(value ?? "")
